requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
faiss-cpu==1.7.4
langchain==0.0.350
langgraph==0.0.20
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
PyPDF2==3.0.1
openpyxl==3.1.2
pandas==2.0.3
python-docx==0.8.11
